refactor: 인기상품 페이지 상품 개수 작업
