+++ src/components/common/Card/CardList/Hot/CardListHot.module.scss
@@ -8,5 +8,29 @@
     grid-template-columns: repeat(2, 1fr);
     row-gap: 2rem;
     column-gap: 1.6rem;
+    place-items: center;
+  }
+
+  .sortBox {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+    width: 100%;
+    padding: 0.8rem 1.6rem 0;
+  }
+
+  .productsCount {
+    @include font-body4;
+
+    color: $color-gray-600;
+    margin: 1.2rem 0;
+    align-self: flex-start;
+  }
+
+  .divider {
+    background-color: $color-gray-50;
+    width: 100%;
+    height: 0.8rem;
+    margin-bottom: 2rem;
   }
 }
