+++ src/pages/products/hot/ProductHotPage.module.scss
@@ -14,22 +14,4 @@
       transform: translateY(-6.8rem);
     }
   }
-
-  .sortBox {
-    display: flex;
-    width: 100%;
-    padding: 0.8rem 1.6rem;
-    justify-content: flex-end;
-  }
-
-  .divider {
-    background-color: $color-gray-50;
-    width: 100%;
-    height: 0.8rem;
-  }
-
-  .contents {
-    min-height: 100vh;
-    padding: 2rem 0;
-  }
 }
