+++ src/pages/products/hot/index.tsx
@@ -7,23 +7,13 @@ import LogoFull from '@/components/common/Icon/LogoFull';
 import SearchButton from '@/components/common/Button/Search';
 import CartButton from '@/components/common/Button/Cart';
 import NavTop from '@/components/common/Nav/Top';
-import GitHubBox from '@/components/common/GitHubBox';
 import FloatingBox from '@/components/common/Layout/Footer/FloatingBox';
 import NavBottom from '@/components/common/Nav/Bottom';
-import SortButton from '@/components/common/Button/Sort';
 import useToast from '@/hooks/useToast';
 import PetToggleButton from '@/components/common/Button/PetToggle';
 import ScrollTopButton from '@/components/common/Button/ScrollTop';
 import CardListHot from '@/components/common/Card/CardList/Hot';
 import FloatingActionBox from '@/components/common/Layout/Footer/FloatingActionBox';
-
-const SORT_OPTIONS = [
-  { name: '최신순', value: '0' },
-  { name: '별점 높은 순', value: '1' },
-  { name: '별점 낮은 순', value: '2' },
-  { name: '가격 높은 순', value: '3' },
-  { name: '가격 낮은 순', value: '4' },
-];
 
 const BOTTOM_BOX_ID = 'bottomBox';
 
@@ -61,26 +51,7 @@ export default function ProductHotPage({ petType, orderBy }: ProductHotPageProps
         </Header.Box>
         <NavTop />
       </Header.Root>
-      <div className={styles.sortBox}>
-        <SortButton
-          options={SORT_OPTIONS}
-          initialOptionValue={orderBy}
-          onClick={value => {
-            router.replace({
-              pathname: '/products/hot',
-              query: {
-                ...router.query,
-                orderBy: value,
-              },
-            });
-          }}
-        />
-      </div>
-      <div className={styles.divider} />
-      <div className={styles.contents}>
-        <CardListHot petType={petType} orderBy={orderBy} />
-      </div>
-      <GitHubBox />
+      <CardListHot petType={petType} orderBy={orderBy} />
       <FloatingBox id={BOTTOM_BOX_ID}>
         <NavBottom />
         <FloatingActionBox>
